Moving theme directory to src
